# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PC 견적 추천 API",
    "version": "2.0.0",
    "description": "PC 부품 조회·검색·비교, 가격 추세·멀티몰 비교, 견적 저장/공유/PDF, 가격 알림, 호환성 검사, AI 추천·업그레이드 어드바이저 API"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "paths": {
...
    "/api/parts": {
      "get": {
        "tags": ["parts"],
        "summary": "부품 목록 조회 (페이징)",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["cpu","gpu","motherboard","memory","storage","psu","case","cooler"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100, "maximum": 200 } }
        ],
        "responses": {
          "200": { "description": "부품 목록 (헤더: X-Total-Count, X-Page, X-Total-Pages)" }
        }
      }
    },
...
    "/api/builds": {
      "post": {
        "tags": ["builds"],
        "summary": "견적 저장 및 공유 링크 생성 (30일 유효)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "builds": [
                  { "category": "cpu", "name": "AMD Ryzen 5 7600X", "price": 250000 },
                  { "category": "gpu", "name": "ASUS TUF GAMING RTX 4070", "price": 550000 }
                ],
                "purpose": "게임용",
                "budget": 1500000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "공유 ID 및 만료일",
            "content": {
              "application/json": {
                "example": { "shareId": "a1b2c3d4", "expiresAt": "2026-05-21T00:00:00.000Z" }
              }
            }
          }
        }
      }
    },
...
    "/api/alerts": {
      "post": {
        "tags": ["alerts"],
        "summary": "가격 알림 등록",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "category": "gpu",
                "name": "ASUS TUF GAMING RTX 4070",
                "targetPrice": 500000,
                "email": "user@example.com"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "알림 등록 완료" },
          "409": { "description": "동일 알림 이미 존재" }
        }
      },
      "get": {
        "tags": ["alerts"],
        "summary": "내 가격 알림 목록 조회",
        "parameters": [
          { "name": "email", "in": "query", "required": true, "schema": { "type": "string", "format": "email" } }
        ],
        "responses": {
          "200": { "description": "알림 목록" }
        }
      }
    },
...
    "/api/recommend": {
      "post": {
        "tags": ["recommend"],
        "summary": "AI 기반 PC 견적 추천 (GPT-4o-mini, 24h 캐시)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": { "budget": 1500000, "purpose": "게임용" }
            }
          }
        },
        "responses": {
          "200": { "description": "추천 견적 목록 + AI 평가" },
          "429": { "description": "요청 횟수 초과 (분당 10회)" }
        }
      }
    },
...
    "/api/gpt-info": {
      "post": {
        "tags": ["parts"],
        "summary": "GPT 부품 한줄평 + 사양 요약 (분당 10회 제한)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": { "partName": "AMD Ryzen 5 7600X" }
            }
          }
        },
        "responses": {
          "200": { "description": "{ review, specSummary }" },
          "429": { "description": "요청 횟수 초과" }
        }
      }
    },
    "/api/naver-price": {
      "get": {
        "tags": ["prices"],
        "summary": "네이버 쇼핑 원시 검색 (직접 쿼리)",
        "parameters": [
          { "name": "query", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "네이버 쇼핑 API 원본 응답" }
        }
      }
    }
  }
}